FORMAT-SQL-RESULTS format SQL results data to make it more readable
